fixed "empty" arrays
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@8537 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/monitor/nvaluebox.h
@@ -201,7 +201,7 @@ class ValueBoxRectangle:public ValueBox, NWindowEdit
 		virtual bool setCursor ();
 	private:
 		NWindowEditIntegers * edt[4];
-		int edtSelected;
+		size_t edtSelected;
 };
 
 /**
@@ -220,7 +220,7 @@ class ValueBoxArray:public ValueBox, NWindowEdit
 		virtual bool setCursor ();
 	private:
 		std::vector <NWindowEdit *> edt;
-		int edtSelected;
+		size_t edtSelected;
 };
 
 /**
